Fix droplets not being destroyed when killing a slave
Digital Ocean API doesn't allow to issue a command on a droplet if
there is some "pending event" going on with the droplet. For example,
if a droplet is being created, you can't destroy it until the creation
is completed. If you try to destroy it before the droplet creation
completes, the DigitalOcean Java API library will throw an exception,
which is not handled when destroying a slave in this plugin. The
exception not being handled results in the plugin mistakenly thinking
that the droplet was actually destroyed, destroying the Jenkins slave
object and removing the slave from Jenkins Nodes list in the UI, while
the droplet associated with the destroyed slave is actually still
running, as the destruction failed.
We want a sure way of destroying a droplet, so the best we can do is if
there is some pending event going on, just keep trying to destroy a
droplet until the destruction succeeds. For that, we introduce a new
method which spawns a thread which tries to destroy a droplet until the
droplet destruction succeeds. The method uses some static state to keep
track of which droplet IDs have already a "destruction thread" created
for them, to prevent several destruction threads spawning for the same
droplet, as it's just a waste of the API rate limit and machine
resources.