Remove bootstrap and extra SSH reconnection from computer launcher
It looks quite pointless, all it does is trying to authenticate using
the same exact credentials in case the last authentication fails. Why
would the second attempt succeed if the first one has failed?

Also removed the SSH reconnecting, followed by the bootstrap. It's
pointless too. Why connect to SSH, authenticate and then disconnect to
connect back again with the same credentials? This just increases the
time it takes to provision a slave.