Fixed droplets not always being destroyed (#24)
The previous droplet destroying code relied on the exception message to contain
certain words, which didn't work in all of the cases, leaving a droplet
undestroyed. It also wasn't very efficient as it was creating a thread per a
droplet we trying to destroy, when having just one thread would suffice.