Add an option for label-less jobs
Previously, if a Slave Template had any labels set, a job not restricted to
build on any node would not trigger creation of a Droplet based on such Slave
Template. In fact, only Slave Templates with no labels specified were handling
jobs not restricted to any label.

This commit adds an option which allows jobs not restricted to any node to
trigger the creation of a Droplet based on a Slave Templates that has labels
set.