log premature return from computer launching
returning because a sub-function failed is all nice and dandy, given
the listener logger is ephemeral, not leaving persistent information as
to why the launch failed is getting in the way of debugging.
the only way to figure out what happened after the fact is looking at the
jenkins log it seems.

to enable finding out why launching failed log what went wrong before
returning