do not attempt to list ssh keys without auth token
when creating a new DO cloud doFillSshKeyIdItems will be called initially,
with an empty authToken. attempting to list the sshkeys then will cause
an exception, which not only looks silly and screws up the config page,
it also seems to confuse stapler/jelly over when to reload the model
(most times for me entering an auth key did not actually trigger a reload)

by only filling the sshkey model if we got a !empty key, reloading seems
overall more consistent and adding a new cloud doesn't greet the user with
an ugly unhandled exception error.

(also changed the iter variable name from image to key :O)