codify the droplet type (regular, high-cpu, high-mem) in the label string (#36)
DO a while ago introduced special droplet variants without cpu stealing or
enormous amounts of ram. They are listed along the regular sizes which
makes them hard to spot in the list. As they have different pricing and
resource behavior than regular nodes it makes sense to more clearly
indicate which is which.