Repeatable template (#39)
* use repeatableProperty to render template config

SlaveTemplate is a describable so we can let repeatableProperty do the
heavy lifting of loading the right config.jelly with the right descriptor
scoping.

Also, move delete button out of the SlaveTemplate config.jelly to
the Cloud's. Defining the button in the describable itself is deprecated.

* enable shuffling of droplets

repeatableProperty can add a handle to re-order repeatables if a header
is defined. as this can be handy to manage a greater number of droplets
to sort them in a more logical/human readable order, set a simple header
to get this feature enabled.