convert old configs to new objects and force sane default values
I am actually not sure how this was dealt with in the past, but currently
the Cloud object is unmarshalled from the config.xml if present, which
means new members are null unless the user goes to the global config and
saves it so that they get properly initialized.
To deal with this we are now adding custom unmarshalling for the new
members to get initialized with sane default values.

This prevents nullptr exceptions when loading the new plugin on old configs