fix instanceCap resetting to default value every time you open /configure
getters need to be public since they get called by jelly render to actually
get the runtime value of the associated entry field.
because getInstanceCap was private the currnet value couldn't be
retrieved and the config view was always initialized to the default value
of 5.

simply make the getter public again.
this broken in some style fixes I did before 0.15.