Added an option allowing private networking inspite of master/slave communication
This option allows one to enable private networking on their slave
regardless of whether or not the master will communicate with it
over the internal network.

This comes in handy for build tasks that will execute on the slaves that
need to communicate to a (static) droplet. For instance, pushing to or
pulling from an internal maven repo.