Deadlock fixes (#47)
* fix deadlock?

* make canprovision not claim the provisionLock

it's super unclear if this must be true or false accurately, but from
looking at other clouds it seems that it can at least not be accurate, so
long as provision then handles that properly. which is the case for us

with the lock being claimed we had deadlocks occurring though I am not sure
what changed that this now happens but previously did not.