optionally allow image resolution by name
this enables a use case where one would continuously update the image
and re-create it as a new snapshot, causing the id to constantly change
but the name to be persistent. since ordinarily name is a really poor
way to resolve images, in particular also because we always need to talk
to the api to get the actual image, this entire feature is limited to
user images rather than all images. specifically this also lessens the
API work we need to do.

the template config has a new checkbox to enable this feature, which
then prompts a reload of the image combobox to only include user images
(and at the same time the values in the model are changed from slugOrIds
to names)

at provisioning time we then need to pass a bit more context along so as
to resolve accordingly in newImage()